refc: transfer fetch function to admin module
