+++ src/api/student/student_route.ts
@@ -3,9 +3,6 @@ import * as studentController from "./student_controller";
 
 const router = Router();
 
-//this is admin request, add to auth module later..
-router.get("/fetch", studentController.fetchUnverifiedStudents);
-
 //handle registration
 router.post("/submit", studentController.studentRegistration);
